Revert "Refactoring Scheduler Package => Changed CancelableScheduler From Int…"
