+++ src/main/java/com/corundumstudio/socketio/scheduler/HashedWheelScheduler.java
@@ -15,23 +15,52 @@
  */
 package com.corundumstudio.socketio.scheduler;
 
+import java.util.Map;
 import java.util.concurrent.ThreadFactory;
 import java.util.concurrent.TimeUnit;
 
 import io.netty.channel.ChannelHandlerContext;
+import io.netty.util.HashedWheelTimer;
 import io.netty.util.Timeout;
 import io.netty.util.TimerTask;
+import io.netty.util.internal.PlatformDependent;
 
-public class HashedWheelScheduler extends CancelableScheduler {
+public class HashedWheelScheduler implements CancelableScheduler {
 
-    private volatile ChannelHandlerContext ctx;
-
+    private final Map<SchedulerKey, Timeout> scheduledFutures = PlatformDependent.newConcurrentHashMap();
+    private final HashedWheelTimer executorService;
+    
     public HashedWheelScheduler() {
-        super();
+        executorService = new HashedWheelTimer();
     }
     
     public HashedWheelScheduler(ThreadFactory threadFactory) {
-        super(threadFactory);
+        executorService = new HashedWheelTimer(threadFactory);
+    }
+
+    private volatile ChannelHandlerContext ctx;
+
+    @Override
+    public void update(ChannelHandlerContext ctx) {
+        this.ctx = ctx;
+    }
+
+    @Override
+    public void cancel(SchedulerKey key) {
+        Timeout timeout = scheduledFutures.remove(key);
+        if (timeout != null) {
+            timeout.cancel();
+        }
+    }
+
+    @Override
+    public void schedule(final Runnable runnable, long delay, TimeUnit unit) {
+        executorService.newTimeout(new TimerTask() {
+            @Override
+            public void run(Timeout timeout) throws Exception {
+                runnable.run();
+            }
+        }, delay, unit);
     }
 
     @Override
@@ -52,7 +81,9 @@ public class HashedWheelScheduler extends CancelableScheduler {
             }
         }, delay, unit);
 
-        addScheduledFuture(key, timeout);
+        if (!timeout.isExpired()) {
+            scheduledFutures.put(key, timeout);
+        }
     }
 
     @Override
@@ -68,7 +99,9 @@ public class HashedWheelScheduler extends CancelableScheduler {
             }
         }, delay, unit);
 
-        addScheduledFuture(key, timeout);
+        if (!timeout.isExpired()) {
+            scheduledFutures.put(key, timeout);
+        }
     }
 
     @Override
@@ -76,10 +109,4 @@ public class HashedWheelScheduler extends CancelableScheduler {
         executorService.stop();
     }
 
-    private void addScheduledFuture(final SchedulerKey key, Timeout timeout){
-        if (!timeout.isExpired()) {
-            scheduledFutures.put(key, timeout);
-        }
-    }
-    
 }
